Update Discord4j to 3.3.1 and refactor YAML parsing logic
- Bump Discord4j dependency version from 3.3.0 to 3.3.1 in `build.gradle.kts`.
- Move `parsePlayerYaml` logic from `GameService` to `RoomController`, consolidating YAML handling.
- Refactor to simplify dependencies and improve code structure.
